+++ lib/src/lib.rs
@@ -15,6 +15,7 @@ pub mod settings;
 pub mod shutdown;
 pub mod utils;
 
+mod accept_limiter;
 mod connection_limiter;
 mod datagram_pipe;
 mod direct_forwarder;